+++ src/pages/Auth.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import LeftPanel from "../components/auth/LeftPanel";
+import LeftPanel from "../components/auth/AuthSidebar";
 import AuthFormFields from "../components/auth/AuthFormFields";
-import { LOGIN_MODE, SIGNUP_MODE} from "../constants/Auth";
+import { LOGIN_MODE, SIGNUP_MODE} from "../constants/auth";
 import type { Field, FormFields } from "../types/form";
 import { AuthContent, AuthMain, AuthWrapper } from "../styles/auth/auth-main";
 import { AuthBrand, AuthDivider, AuthFooter, AuthLine } from "../styles/auth/auth-main";
