+++ src/pages/Login.tsx
@@ -1,5 +1,5 @@
 import AuthPage from "./Auth";
-import { LOGIN_MODE } from "../constants/Auth";
+import { LOGIN_MODE } from "../constants/auth";
 
 const Login = () => {
   return <AuthPage mode={LOGIN_MODE} />;
