+++ src/types/auth.ts
@@ -1,4 +1,4 @@
-import { LOGIN_MODE, SIGNUP_MODE } from "../constants/Auth";
+import { LOGIN_MODE, SIGNUP_MODE } from "../constants/auth";
 import type { Field, FormFields } from "./form"
 
 export type Mode = typeof LOGIN_MODE | typeof SIGNUP_MODE;
