+++ src/types/column.ts
@@ -1,0 +1,8 @@
+import type { ColumnItem } from "./board-view";
+
+export type ColumnProps = {
+  column: ColumnItem;
+  onRename: (id: string, newTitle: string) => void;
+  onDelete: (id: string) => void;
+  boardId: string;
+};
