+++ src/utils/validation.ts
@@ -5,24 +5,27 @@ export function validatePassword(password: string): string | null {
   const digitPattern = /\d/;
   const specialCharPattern = /[!@#$%^&*(),.?":{}|<>]/;
 
-  if (password.length < minLength) {
-    return "Password must be at least 8 characters long.";
-  }
-  if (!uppercasePattern.test(password)) {
-    return "Password must contain at least one uppercase letter.";
-  }
-  if (!lowercasePattern.test(password)) {
-    return "Password must contain at least one lowercase letter.";
-  }
-  if (!digitPattern.test(password)) {
-    return "Password must contain at least one digit.";
-  }
-  if (!specialCharPattern.test(password)) {
-    return "Password must contain at least one special character.";
-  }
+  switch (true) {
+    case password.length < minLength:
+      return "Password must be at least 8 characters long.";
 
-  return null;
+    case !uppercasePattern.test(password):
+      return "Password must contain at least one uppercase letter.";
+
+    case !lowercasePattern.test(password):
+      return "Password must contain at least one lowercase letter.";
+
+    case !digitPattern.test(password):
+      return "Password must contain at least one digit.";
+
+    case !specialCharPattern.test(password):
+      return "Password must contain at least one special character.";
+
+    default:
+      return null;
+  }
 }
+
 
 export function validateEmail(email: string): string | null {
   const emailPattern = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
